backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
akshare==1.18.22
anaconda-anon-usage @ file:///opt/miniconda3/conda-bld/anaconda-anon-usage_1764636648062/work
anaconda-auth @ file:///opt/miniconda3/conda-bld/anaconda-cloud-auth-split_1765839823423/work
anaconda-cli-base @ file:///opt/miniconda3/conda-bld/anaconda-cli-base_1764888589923/work
annotated-doc==0.0.4
annotated-types @ file:///opt/miniconda3/conda-bld/annotated-types_1761744949290/work
anyio==4.12.1
APScheduler==3.11.2
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
attrs==25.4.0
baostock==0.8.9
beautifulsoup4==4.14.3
boltons @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_97h34kncb9/croot/boltons_1751383736162/work
brotlicffi @ file:///opt/miniconda3/conda-bld/brotlicffi_1764961161081/work
certifi @ file:///opt/miniconda3/conda-bld/certifi_1763135329285/work/certifi
cffi @ file:///opt/miniconda3/conda-bld/cffi_1761832773185/work
charset-normalizer @ file:///opt/miniconda3/conda-bld/charset-normalizer_1761744840553/work
click @ file:///opt/miniconda3/conda-bld/click_1764332239180/work
colorama==0.4.6
conda @ file:///opt/miniconda3/conda-bld/conda_1765567809749/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ebkcs5yxh6/croot/conda-anaconda-tos_1755123330773/work
conda-content-trust @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-content-trust_1728603284288/work
conda-libmamba-solver @ file:///C:/miniconda3/conda-bld/conda-libmamba-solver_1764245615235/work/src
conda-package-handling @ file:///opt/miniconda3/conda-bld/conda-package-handling_1762366402479/work
conda_package_streaming @ file:///opt/miniconda3/conda-bld/conda-package-streaming_1762361629849/work
contourpy==1.3.3
coverage==7.13.3
cryptography @ file:///opt/miniconda3/conda-bld/cryptography-split_1761931964095/work
curl_cffi==0.14.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/distro_1728591855492/work
et_xmlfile==2.0.0
faiss-cpu==1.13.2
fastapi==0.128.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozendict @ file:///opt/miniconda3/conda-bld/frozendict_1761750665916/work
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna @ file:///opt/miniconda3/conda-bld/idna_1761912005426/work
iniconfig==2.3.0
jaraco.classes @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_97dr4u6nc1/croot/jaraco.classes_1755516335437/work
jaraco.context @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jaraco.context_1731716684676/work
jaraco.functools @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_91zaq5_008/croot/jaraco.functools_1740408121405/work
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jsonpatch_1728594556110/work
jsonpath==0.82.2
jsonpointer @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_bdid71leqg/croot/jsonpointer_1753788445166/work
keyring @ file:///opt/miniconda3/conda-bld/keyring_1763637309871/work
kiwisolver==1.4.9
langchain==1.2.9
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.9
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
lark-oapi==1.5.3
libmambapy @ file:///opt/miniconda3/conda-bld/mamba-split_1763111381025/work/libmambapy
loguru==0.7.3
lxml==6.0.2
markdown-it-py @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6av49xs_3l/croot/markdown-it-py_1756299201304/work
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl @ file:///opt/miniconda3/conda-bld/mdurl_1758552175942/work
menuinst @ file:///opt/miniconda3/conda-bld/menuinst_1765382369523/work
mini-racer==0.14.1
ml_dtypes==0.5.4
modelscope==1.34.0
more-itertools @ file:///opt/miniconda3/conda-bld/more-itertools_1761121541654/work
mpmath==1.3.0
msgpack @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0d73gp34a_/croot/msgpack-python_1750958625811/work
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.2
onnx==1.20.1
onnx-ir==0.1.15
onnxoptimizer==0.4.2
onnxruntime==1.24.1
onnxscript==0.6.0
openai==2.17.0
openpyxl==3.1.5
orjson==3.11.7
ormsgpack==1.12.2
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079023/work
pandas==3.0.0
pillow==12.1.0
pkce @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pkce_1728592680499/work
platformdirs @ file:///opt/miniconda3/conda-bld/platformdirs_1762356666077/work
pluggy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_bei6q1cvvt/croot/pluggy_1733169621717/work
propcache==0.4.1
protobuf==6.33.5
pycosat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_61k1o_oje7/croot/pycosat_1736868429741/work
pycparser @ file:///opt/miniconda3/conda-bld/pycparser_1757496048647/work
pycryptodome==3.23.0
pydantic @ file:///opt/miniconda3/conda-bld/pydantic_1764083527102/work
pydantic-settings @ file:///opt/miniconda3/conda-bld/pydantic-settings_1764165151428/work
pydantic_core @ file:///opt/miniconda3/conda-bld/pydantic-core_1764009784214/work
Pygments @ file:///opt/miniconda3/conda-bld/pygments_1762431584667/work
PyJWT @ file:///opt/miniconda3/conda-bld/pyjwt_1764332099092/work
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
pyqtgraph==0.14.0
PySocks @ file:///opt/miniconda3/conda-bld/pysocks_1761753010927/work
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-qt==4.5.0
python-dateutil==2.9.0.post0
python-dotenv @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0bqseadoho/croot/python-dotenv_1745610246418/work
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
readchar @ file:///opt/miniconda3/conda-bld/readchar_1760613300945/work
regex==2026.1.15
reportlab==4.4.9
requests @ file:///opt/miniconda3/conda-bld/requests_1762359592010/work
requests-toolbelt==1.0.0
rich @ file:///opt/miniconda3/conda-bld/rich_1760375480293/work
ruamel.yaml @ file:///opt/miniconda3/conda-bld/ruamel.yaml_1762535951282/work
ruamel.yaml.clib @ file:///opt/miniconda3/conda-bld/ruamel.yaml.clib_1762530070208/work
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
semver @ file:///opt/miniconda3/conda-bld/semver_1761903026306/work
setuptools==80.9.0
shellingham @ file:///opt/miniconda3/conda-bld/shellingham_1761912204150/work
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tomli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03b1up8uh3/croot/tomli_1753774598375/work
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm @ file:///opt/miniconda3/conda-bld/tqdm_1762863348900/work
truststore @ file:///opt/miniconda3/conda-bld/truststore_1762520923820/work
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection @ file:///opt/miniconda3/conda-bld/typing-inspection_1760614169744/work
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_dffmcwqo6o/croot/typing_extensions_1756281471323/work
tzlocal==5.3.1
urllib3 @ file:///opt/miniconda3/conda-bld/urllib3_1765399201892/work
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wheel==0.45.1
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zstandard @ file:///opt/miniconda3/conda-bld/zstandard_1758189078532/work
